gaussian_splatting: extract layout-hint subsystem into its own TU (decomposition slice 1)
gaussian_streaming.cpp (6051 LOC) carried a self-contained ~460-LOC "layout
hint" subsystem inside its anonymous namespace: the failure-tracking types
(LayoutHintUsage/FailureReason/FailureCategory/ValidationFailure/
FailureCounters/FailureState/IndexRange), the per-system failure-state map,
the _layout_hint_* reporting helpers, and _validate_layout_hint_ranges. This
cluster is used by 6+ method groups (ctor/dtor, both init paths, chunk
building, end_frame, the diagnostics snapshot) — i.e. genuinely shared
infrastructure that does not belong buried in the god-file's internal
linkage.

Move it verbatim (zero logic change) into a new translation unit
core/streaming_layout_hint.{h,cpp} under namespace gs_layout_hint.
gaussian_streaming.cpp now includes the header and adds a file-scope
`using namespace gs_layout_hint;` so all ~30 existing call sites compile
unchanged. SCsub/CMake glob core/*.cpp, so no build-script edit is needed.

Result: gaussian_streaming.cpp shrinks 6051 -> 5592 LOC; the layout-hint
subsystem becomes an independently reviewable 109-LOC header + 407-LOC unit.
This is behavior-preserving and unblocks the next slice (the diagnostics
surface, which depends on _layout_hint_build_snapshot).

Risk: R2 (streaming). Evidence: incremental editor build compiles + links
clean (scons platform=windows target=editor dev_build=yes tests=yes, exit 0);
guards green (run_module_tests.py --guard-only, 108 tests OK).
